Move List and ListItem to common
