+++ packages/onboarding-ui/src/common/List/index.ts
@@ -1,0 +1,6 @@
+import List from './List';
+import ListItem from './ListItem';
+
+export default Object.assign(List, {
+  Li: ListItem,
+});
